frontend: move lib/components/saft to saft/_components
