+++ frontend/src/routes/saft/_components/MessageAfterwards.svelte
