+++ frontend/src/routes/saft/_components/MessageClosed.svelte
